Added required data/lines directory
lab/data/lines contained all the atomic data for the line emission and 2-photon.  It was left out because of a faulty .gitignore.  Edited gitignore and removed references to map51. Added to main brance as it is a code breaker otherwise, will stick to rss edits otherwise.  Will test on a completely fresh user install tonight.  RSS