Fixed build fail on a completely fresh clone, fixed tools/
Fixed tools directory for line fitting anf convolution of mappings spectra as well as reddening and de-reddening observed as well as mappings files. Updated gitignoe to only ignore the build binaries.  Still neeed to fix main data files.  Recommend GIT-LFS as it is transparent to the end users and us once it is installed and they use the github app to setup.
RSS